add push workflow
add aws to template-manager

update template-manager

+++ packages/template-manager/internal/template/storage.go
@@ -4,25 +4,52 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/e2b-dev/infra/packages/shared/pkg/consts"
 	"github.com/e2b-dev/infra/packages/shared/pkg/storage"
 	"github.com/e2b-dev/infra/packages/shared/pkg/storage/gcs"
+	"github.com/e2b-dev/infra/packages/shared/pkg/storage/s3"
 )
 
 type Storage struct {
-	bucket *gcs.BucketHandle
+	bucket   *gcs.BucketHandle
+	s3bucket *s3.BucketHandle
 }
 
 func NewStorage(ctx context.Context) *Storage {
-	return &Storage{
-		bucket: gcs.GetTemplateBucket(),
+	if consts.CloudProviderEnv == consts.AWS {
+		return &Storage{
+			bucket:   nil,
+			s3bucket: s3.GetTemplateBucket(),
+		}
 	}
+
+	if consts.CloudProviderEnv == consts.GCP {
+		return &Storage{
+			bucket:   gcs.GetTemplateBucket(),
+			s3bucket: nil,
+		}
+	}
+
+	panic(fmt.Sprintf("invalid cloud provider: %s", consts.CloudProviderEnv))
 }
 
 func (t *Storage) Remove(ctx context.Context, buildId string) error {
-	err := gcs.RemoveDir(ctx, t.bucket, buildId)
-	if err != nil {
-		return fmt.Errorf("error when removing template '%s': %w", buildId, err)
+
+	if consts.CloudProviderEnv == consts.AWS {
+		err := s3.RemoveDir(ctx, t.s3bucket, buildId)
+		if err != nil {
+			return fmt.Errorf("error when removing template '%s': %w", buildId, err)
+		}
 	}
+
+	if consts.CloudProviderEnv == consts.GCP {
+		err := gcs.RemoveDir(ctx, t.bucket, buildId)
+		if err != nil {
+			return fmt.Errorf("error when removing template '%s': %w", buildId, err)
+		}
+	}
+
+	panic(fmt.Sprintf("invalid cloud provider: %s", consts.CloudProviderEnv))
 
 	return nil
 }
